refactor(capture): extract virtual feature records
